requirements.txt
streamlit
pandas
pymupdf
requests
supabase-py